requirements.txt
# Core
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# ANN Search
faiss-cpu>=1.7.4

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# LLM Reranker
openai>=1.0.0

# Utilities
tqdm>=4.65.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
